+++ packages/react-markdown/src/index.ts
@@ -9,16 +9,4 @@ export type {
 } from "./overrides/types";
 
 export { useIsMarkdownCodeBlock } from "./overrides/PreOverride";
-
-export {
-  /** @deprecated This export was moved to `@assistant-ui/react-ui`. Use the migration codemod to update your code, by running `npx assistant-ui upgrade`. */
-  makeMarkdownText,
-
-  /** @deprecated This export was moved to `@assistant-ui/react-ui`. Use the migration codemod to update your code, by running `npx assistant-ui upgrade`. */
-  type MakeMarkdownTextProps,
-} from "./ui/markdown-text";
-
-/** @deprecated This export was moved to `@assistant-ui/react-ui`. Use the migration codemod to update your code, by running `npx assistant-ui upgrade`. */
-export { CodeHeader } from "./ui/code-header";
-
 export { memoizeMarkdownComponents as unstable_memoizeMarkdownComponents } from "./memoization";