+++ packages/react/src/primitives/contentPart/index.ts
@@ -1,4 +1,3 @@
 export { ContentPartPrimitiveText as Text } from "./ContentPartText";
 export { ContentPartPrimitiveImage as Image } from "./ContentPartImage";
-export { ContentPartPrimitiveDisplay as Display } from "./ContentPartDisplay";
 export { ContentPartPrimitiveInProgress as InProgress } from "./ContentPartInProgress";