+++ packages/react/src/runtimes/edge/converters/toLanguageModelMessages.ts
@@ -133,7 +133,7 @@ export function toLanguageModelMessages(
                 }
 
                 default: {
-                  const unhandledType: "ui" | "audio" = type;
+                  const unhandledType: "audio" = type;
                   throw new Error(
                     `Unspported content part type: ${unhandledType}`,
                   );
@@ -163,7 +163,7 @@ export function toLanguageModelMessages(
               break;
             }
             default: {
-              const unhandledType: "ui" = type;
+              const unhandledType: never = type;
               throw new Error(`Unhandled content part type: ${unhandledType}`);
             }
           }